requirements.txt
ultralytics>=8.0.0
Pillow>=9.0.0
beautifulsoup4>=4.11.0
numpy>=1.21.0
requests>=2.28.0
supervision>=0.16.0
opencv-python>=4.7.0
ipython>=8.0.0